+++ com.unity.netcode.gameobjects/CHANGELOG.md
@@ -19,6 +19,7 @@ Additional documentation and release notes are available at [Multiplayer Documen
 
 - Updated `UnityTransport` dependency on `com.unity.transport` to 1.1.0. (#2025)
 - (API Breaking) `ConnectionApprovalCallback` is no longer an `event` and will not allow more than 1 handler registered at a time. Also, `ConnectionApprovalCallback` is now a `Func<>` taking `ConnectionApprovalRequest` in and returning `ConnectionApprovalResponse` back out (#1972)
+- (API Breaking) `NetworkTransform.OnIsServerAuthoritatitive` is now obsolete. Use `NetworkTransform.OnIsServerAuthoritative` instead.
 
 ### Removed